requirements.txt
langchain
ipykernel
python-dotenv
langchain_community
pypdf
bs4
arxiv
pymupdf
langchain-text-splitters
chromadb
langchain-ollama
sentence_transformers
langchain_huggingface
faiss-cpu
langchain_chroma
fastapi
langserve
sentence-transformers
wikipedia
arxiv
ddgs
duckduckgo-search